pyproject.toml
[project]
name = "quant-trading-supportive-system"
version = "1.0.0"
description = "Private cryptocurrency analysis platform"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0.0,<5.0.0",
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "duckdb>=0.9.0,<1.0.0",
    "pyarrow>=14.0.0,<15.0.0",
    "streamlit>=1.28.0,<2.0.0",
    "streamlit-lightweight-charts>=0.8.0,<1.0.0",
    "pandas-ta>=0.3.14b,<1.0.0",
    "ta>=0.10.0,<1.0.0",
    "stumpy>=1.12.0,<2.0.0",
    "tslearn>=0.6.0,<1.0.0",
    "vectorbt>=0.26.0,<1.0.0",
    "backtesting>=0.3.3,<1.0.0",
    "prefect>=3.0.0,<4.0.0",
    "websockets>=12.0,<13.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "loguru>=0.7.0,<1.0.0",
    "httpx>=0.25.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pip-audit>=2.6.0",
    "pdoc>=14.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501", "B008", "B905"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-ra -q --strict-markers --cov-branch"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 70
